Houve modificação na tabela produto e a incersão da tabela cor e tamanha. Foi acrescentado mais uma tabela que faz a relação de produto com tamanho e cor, e o Dao de todas as outras entidades foram atualizados.

+++ dao/AdminDao.php
@@ -40,13 +40,28 @@
                 }
         }
 
-        public function remover(string $id):void
+        public function podeLoja(string $id):bool
         {
-            
-            $resultado = $this->mysql->prepare
-            ("delete from tb_admin where idtb_admin = ?");
-            $resultado->bind_param('s', $id);
+            $resultado = $this->mysql->prepare("select * from tb_loja where tb_admin_idtb_admin = ?");
+            $resultado->bind_param('s',$id);
             $resultado->execute();
+            if($resultado->get_result()->num_rows > 0)
+                return false; // não pode
+            else
+                return true; // pode
+        }
+        public function remover(string $id):bool
+        {
+            if ($this->podeLoja($id)){
+                $resultado = $this->mysql->prepare
+                ("delete from tb_admin where idtb_admin = ?");
+                $resultado->bind_param('s', $id);
+                $resultado->execute();
+                return true;
+            }
+            else {
+                return false;
+            }
        
         }
